+++ lib/rubocop/cop/style/empty_line_between_defs.rb
@@ -9,20 +9,30 @@ module RuboCop
         MSG = 'Use empty lines between defs.'
 
         def on_def(node)
-          if @prev_def_end && (def_start(node) - @prev_def_end) == 1
-            unless @prev_was_single_line && singe_line_def?(node) &&
+          if @prev_def_end && @prev_def_end < def_end(node) &&
+             nothing_or_comments_between_end_and_def?(node)
+            unless @prev_was_single_line && single_line_def?(node) &&
                    cop_config['AllowAdjacentOneLineDefs']
               add_offense(node, :keyword)
             end
           end
 
           @prev_def_end = def_end(node)
-          @prev_was_single_line = singe_line_def?(node)
+          @prev_was_single_line = single_line_def?(node)
         end
 
         private
 
-        def singe_line_def?(node)
+        def nothing_or_comments_between_end_and_def?(node)
+          distance = def_start(node) - @prev_def_end
+          return true if distance == 1
+
+          (@prev_def_end + 1...def_start(node)).all? do |line|
+            processed_source.comments.find { |c| c.loc.line == line }
+          end
+        end
+
+        def single_line_def?(node)
           def_start(node) == def_end(node)
         end
 
@@ -35,8 +45,14 @@ module RuboCop
         end
 
         def autocorrect(node)
-          range = range_with_surrounding_space(node.loc.expression, :left)
-          ->(corrector) { corrector.insert_before(range, "\n") }
+          prev_def = node.ancestors.first.children[node.sibling_index - 1]
+          end_pos = prev_def.loc.end.end_pos
+          source_buffer = prev_def.loc.end.source_buffer
+          newline_pos = source_buffer.source.index("\n", end_pos)
+          newline = Parser::Source::Range.new(source_buffer,
+                                              newline_pos,
+                                              newline_pos + 1)
+          ->(corrector) { corrector.insert_after(newline, "\n") }
         end
       end
     end